+++ lib/src/ui/components/index.dart
@@ -1,0 +1,2 @@
+export 'badge.dart';
+export 'image.dart';